Changes the SetOutfit verb to be a component verb instead of a global verb
