+++ Content.Client/GameObjects/Components/HUD/Inventory/ClientInventoryComponent.cs
@@ -4,12 +4,15 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using Content.Client.GameObjects.Components.Clothing;
 using Content.Shared.GameObjects.Components.Inventory;
+using Content.Shared.GameObjects.Verbs;
 using Content.Shared.Preferences.Appearance;
+using Robust.Client.Console;
 using Robust.Client.GameObjects;
 using Robust.Client.Interfaces.GameObjects.Components;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Interfaces.GameObjects;
 using Robust.Shared.IoC;
+using Robust.Shared.Localization;
 using Robust.Shared.ViewVariables;
 using static Content.Shared.GameObjects.Components.Inventory.EquipmentSlotDefines;
 using static Content.Shared.GameObjects.Components.Inventory.SharedInventoryComponent.ClientInventoryMessage;
@@ -239,6 +242,43 @@ namespace Content.Client.GameObjects.Components.HUD.Inventory
             }
 
             return false;
+        }
+
+        [Verb]
+        private sealed class SetOutfitVerb : Verb<ClientInventoryComponent>
+        {
+            public override bool RequireInteractionRange => false;
+            public override bool BlockedByContainers => false;
+
+            protected override void GetData(IEntity user, ClientInventoryComponent component, VerbData data)
+            {
+                data.Visibility = VerbVisibility.Invisible;
+                var target = component.Owner;
+
+                if (!CanCommand()) //|| !target.HasComponent<ClientInventoryComponent>())
+                    return;
+
+                data.Visibility = VerbVisibility.Visible;
+                data.Text = Loc.GetString("Set Outfit");
+                data.CategoryData = VerbCategories.Debug;
+            }
+
+            protected override void Activate(IEntity user, ClientInventoryComponent component)
+            {
+                if (!CanCommand())
+                    return;
+                var console = IoCManager.Resolve<IClientConsole>();
+                var target = component.Owner;
+
+                var entityId = target.Uid.ToString();
+                console.ProcessCommand($"setoutfit {entityId}");
+            }
+
+            private static bool CanCommand()
+            {
+                var groupController = IoCManager.Resolve<IClientConGroupController>();
+                return groupController.CanCommand("setoutfit");
+            }
         }
     }
 }
